+++ src/task/mod.rs
@@ -34,8 +34,9 @@ pub use self::rm_default_attrs::remove_default_attributes;
 pub use self::rm_dupl_defs::{
     remove_dupl_linear_gradients,
     remove_dupl_radial_gradients,
-    remove_dupl_fe_gaussian_blur
+    remove_dupl_fe_gaussian_blur,
 };
+pub use self::rm_invalid_stops::remove_invalid_stops;
 pub use self::rm_invisible_elements::remove_invisible_elements;
 pub use self::rm_needless_attrs::remove_needless_attributes;
 pub use self::rm_text_attrs::remove_text_attributes;
@@ -60,6 +61,7 @@ mod resolve_attrs;
 mod resolve_inherit;
 mod rm_default_attrs;
 mod rm_dupl_defs;
+mod rm_invalid_stops;
 mod rm_invisible_elements;
 mod rm_needless_attrs;
 mod rm_text_attrs;